Arduino TFT display test
Arduino code performs full screen RGB test then scans each line of the display while displaying the line position on screen.

This was created because I noticed the cheap TFT displays from China don't always provide use of the entire screen. This code will show the origin and bounds of a display. Attach the TFT as directed for your board, and attach a potentiometer to analog pin 6. The potentiometer will adjust the scan speed.